fix long to int32_t

+++ esphome/components/binary_storage/flash_partition.cpp
@@ -236,7 +236,7 @@ storage::StorageError FlashPartition::seek(storage::FileHandle *handle, int64_t 
   } else if (mode == storage::SeekMode::END) {
     whence = SEEK_END;
   }
-  if (fseek(handle->file, static_cast<long>(offset), whence) != 0)
+  if (fseek(handle->file, static_cast<int32_t>(offset), whence) != 0)
     return storage::StorageError::INVALID_ARGS;
 
   return storage::StorageError::OK;
@@ -246,7 +246,7 @@ storage::StorageError FlashPartition::tell(storage::FileHandle *handle, uint64_t
   if (handle == nullptr || !handle->in_use || handle->file == nullptr || position == nullptr)
     return storage::StorageError::INVALID_ARGS;
 
-  long pos = ftell(handle->file);
+  int32_t pos = ftell(handle->file);
   if (pos < 0)
     return storage::StorageError::READ_ERROR;
